water-demand-forecast-main/requirements.txt
pandas
scikit-learn
requests
python-dotenv
streamlit
joblib
matplotlib
numpy
python-dateutil
